partial refactor |mailmsg|
good so far, but needs more work for full block structure
