+++ src/libuc/mailmsg.h
@@ -11,20 +11,16 @@
 /* Copyright © 1998 David A­D­ Morano.  All rights reserved. */
 
 #ifndef	MAILMSG_INCLUDE
-#define	MAILMSG_INCLUDE		1
+#define	MAILMSG_INCLUDE
 
 
 #include	<envstandards.h>	/* MUST be first to configure */
-
 #include	<sys/types.h>
 #include	<unistd.h>
-
 #include	<vecobj.h>
 #include	<strpack.h>
 #include	<localmisc.h>
 
-
-/* object defines */
 
 #define	MAILMSG_MAGIC		0x97634587
 #define	MAILMSG			struct mailmsg_head
@@ -54,18 +50,18 @@ typedef MAILMSG		mailmsg ;
 
 EXTERNC_begin
 
-extern int mailmsg_start(MAILMSG *) noex ;
-extern int mailmsg_loadline(MAILMSG *,cchar *,int) noex ;
-extern int mailmsg_envcount(MAILMSG *) noex ;
-extern int mailmsg_envaddress(MAILMSG *,int,cchar **) noex ;
-extern int mailmsg_envdate(MAILMSG *,int,cchar **) noex ;
-extern int mailmsg_envremote(MAILMSG *,int,cchar **) noex ;
-extern int mailmsg_hdrcount(MAILMSG *,cchar *) noex ;
-extern int mailmsg_hdrikey(MAILMSG *,int,cchar **) noex ;
-extern int mailmsg_hdriline(MAILMSG *,cchar *,int,int,cchar **) noex ;
-extern int mailmsg_hdrival(MAILMSG *,cchar *,int,cchar **) noex ;
-extern int mailmsg_hdrval(MAILMSG *,cchar *,cchar **) noex ;
-extern int mailmsg_finish(MAILMSG *) noex ;
+extern int mailmsg_start(mailmsg *) noex ;
+extern int mailmsg_loadline(mailmsg *,cchar *,int) noex ;
+extern int mailmsg_envcount(mailmsg *) noex ;
+extern int mailmsg_envaddress(mailmsg *,int,cchar **) noex ;
+extern int mailmsg_envdate(mailmsg *,int,cchar **) noex ;
+extern int mailmsg_envremote(mailmsg *,int,cchar **) noex ;
+extern int mailmsg_hdrcount(mailmsg *,cchar *) noex ;
+extern int mailmsg_hdrikey(mailmsg *,int,cchar **) noex ;
+extern int mailmsg_hdriline(mailmsg *,cchar *,int,int,cchar **) noex ;
+extern int mailmsg_hdrival(mailmsg *,cchar *,int,cchar **) noex ;
+extern int mailmsg_hdrval(mailmsg *,cchar *,cchar **) noex ;
+extern int mailmsg_finish(mailmsg *) noex ;
 
 EXTERNC_end
